moved style to theme
