+++ packages/wix-ui-core/src/components/image/image.meta.ts
@@ -1,11 +1,11 @@
 import { Image, ImageStatus } from './image';
 import Registry from '@ui-autotools/registry';
-import style from '../../../stories/image/style.st.css';
+import style from '../../themes/default/image/style.st.css';
 import { SRC, BROKEN_SRC, ERROR_IMAGE_SRC } from './test-fixtures';
 
 const imageMetadata = Registry.getComponentMetadata(Image);
 
-imageMetadata.addStyle(style, { name: 'style', path: 'stories/image/style.st.css' });
+imageMetadata.addStyle(style, { name: 'style', path: 'themes/default/image/style.st.css' });
 
 imageMetadata
   .addSim({
