+++ packages/wix-ui-core/src/themes/default/image/style.st.css
@@ -1,5 +1,5 @@
 :import {
-    -st-from: "../../src/components/image/image.st.css";
+    -st-from: "../../../components/image/image.st.css";
     -st-default: Image;
 }
